BUG: Image data format bug.  BUG: Input preprocessing is required.

+++ engine/InputDataProducer.py
@@ -18,7 +18,7 @@ logger = logging.getLogger("InputDataProducer")
 
 def getData(image, xSize, ySize):
     pixel = image.load()
-    data = numpy.zeros((1, ySize, xSize, 3))
+    data = numpy.zeros((1, ySize, xSize, 3), dtype=numpy.uint8)
     for x in range(xSize):
         for y in range(ySize):
             (red,green,blue) = pixel[x,y]
@@ -26,6 +26,7 @@ def getData(image, xSize, ySize):
             data[0, y, x, 1] = green
             data[0, y, x, 2] = blue
 
+    #Image.fromarray(numpy.asarray(numpy.reshape(data, (ySize, xSize, 3)))).show()
     return data
 
 class InputDataProducer:
